+++ python/ray/experimental/gpu_object_manager/gpu_object_manager.py
@@ -3,108 +3,11 @@ from typing import TYPE_CHECKING, Any, Dict, List, NamedTuple, Optional, Tuple
 import ray
 from ray._private.custom_types import TensorTransportEnum
 from ray._raylet import ObjectRef
-import ray.util.collective as collective
-from ray.util.collective.types import Backend
 
-
-try:
-    import torch
-except ImportError:
-    raise ImportError(
-        "`tensor_transport` requires PyTorch. "
-        "Please install torch with 'pip install torch' to use this feature."
-    )
 
 if TYPE_CHECKING:
     import torch
-
-TENSOR_TRANSPORT_TO_COLLECTIVE_BACKEND = {
-    TensorTransportEnum.NCCL: Backend.NCCL,
-    TensorTransportEnum.GLOO: Backend.TORCH_GLOO,
-}
-
-COLLECTIVE_BACKEND_TO_TORCH_DEVICE = {
-    Backend.NCCL: torch.device("cuda"),
-    Backend.TORCH_GLOO: torch.device("cpu"),
-}
-
-
-def _tensor_transport_to_collective_backend(
-    tensor_transport: TensorTransportEnum,
-) -> Backend:
-    try:
-        return TENSOR_TRANSPORT_TO_COLLECTIVE_BACKEND[tensor_transport]
-    except KeyError:
-        raise ValueError(
-            f"Invalid tensor transport {tensor_transport.name}, must be one of {list(TENSOR_TRANSPORT_TO_COLLECTIVE_BACKEND.keys())}."
-        )
-
-
-def __ray_send__(self, communicator_name: str, obj_id: str, dst_rank: int):
-    """Helper function that runs on the src actor to send tensors to the dst actor."""
-    from ray._private.worker import global_worker
-
-    gpu_object_manager = global_worker.gpu_object_manager
-    assert gpu_object_manager.has_gpu_object(
-        obj_id
-    ), f"obj_id={obj_id} not found in GPU object store"
-    tensors = gpu_object_manager.get_gpu_object(obj_id)
-
-    backend = collective.get_group_handle(communicator_name).backend()
-    device = COLLECTIVE_BACKEND_TO_TORCH_DEVICE[backend]
-
-    for tensor in tensors:
-        if tensor.device.type != device.type:
-            # TODO(swang): Right now there is no way to catch this error
-            # and the receiving Ray task will hang.
-            raise ValueError(
-                f"tensor device {tensor.device} does not match device {device}"
-            )
-        collective.send(tensor, dst_rank, group_name=communicator_name)
-    # TODO(kevin85421): The current garbage collection implementation for the
-    # in-actor object store is naive. We garbage collect each object after it
-    # is consumed once.
-    gpu_object_manager.remove_gpu_object(obj_id)
-
-
-def __ray_recv__(
-    self,
-    communicator_name: str,
-    obj_id: str,
-    src_rank: int,
-    tensor_meta: List[Tuple["torch.Size", "torch.dtype"]],
-):
-    """Helper function that runs on the dst actor to receive tensors from the src actor."""
-    from ray._private.worker import global_worker
-
-    backend = collective.get_group_handle(communicator_name).backend()
-    device = COLLECTIVE_BACKEND_TO_TORCH_DEVICE[backend]
-
-    gpu_object_manager = global_worker.gpu_object_manager
-    tensors = []
-    for meta in tensor_meta:
-        shape, dtype = meta
-        tensor = torch.zeros(shape, dtype=dtype, device=device)
-        collective.recv(tensor, src_rank, group_name=communicator_name)
-        tensors.append(tensor)
-    gpu_object_manager.add_gpu_object(obj_id, tensors)
-
-
-def __ray_fetch_gpu_object__(self, obj_id: str):
-    """Helper function that runs on the src actor to fetch tensors from the GPU object store via the object store."""
-    from ray._private.worker import global_worker
-
-    gpu_object_manager = global_worker.gpu_object_manager
-    assert gpu_object_manager.has_gpu_object(
-        obj_id
-    ), f"obj_id={obj_id} not found in GPU object store"
-    tensors = gpu_object_manager.get_gpu_object(obj_id)
-    # TODO(kevin85421): The current garbage collection implementation for the
-    # in-actor object store is naive. We garbage collect each object after it
-    # is consumed once.
-    gpu_object_manager.remove_gpu_object(obj_id)
-    return tensors
-
+    from ray.experimental.gpu_object_manager.gpu_object_store import GPUObjectStore
 
 # GPUObjectMeta is a named tuple containing the source actor, tensor transport
 # backend, and tensor metadata.
@@ -120,26 +23,52 @@ class GPUObjectMeta(NamedTuple):
     tensor_meta: List[Tuple["torch.Size", "torch.dtype"]]
 
 
+def __ray_get_tensor_meta__(self, obj_id: str):
+    """Helper function that runs on the src actor to get the tensor metadata."""
+    from ray._private.worker import global_worker
+
+    gpu_object_store = global_worker.gpu_object_manager.gpu_object_store
+    assert gpu_object_store.has_gpu_object(
+        obj_id
+    ), f"obj_id={obj_id} not found in GPU object store"
+    tensors = gpu_object_store.get_gpu_object(obj_id)
+    return [(t.shape, t.dtype) for t in tensors]
+
+
+def __ray_fetch_gpu_object__(self, obj_id: str):
+    """Helper function that runs on the src actor to fetch tensors from the GPU object store via the object store."""
+    from ray._private.worker import global_worker
+
+    gpu_object_store = global_worker.gpu_object_manager.gpu_object_store
+    assert gpu_object_store.has_gpu_object(
+        obj_id
+    ), f"obj_id={obj_id} not found in GPU object store"
+    tensors = gpu_object_store.get_gpu_object(obj_id)
+    # TODO(kevin85421): The current garbage collection implementation for the
+    # in-actor object store is naive. We garbage collect each object after it
+    # is consumed once.
+    gpu_object_store.remove_gpu_object(obj_id)
+    return tensors
+
+
 class GPUObjectManager:
     def __init__(self):
-        # A dictionary that maps from an object ID to a list of tensors.
-        #
-        # Note: Currently, `gpu_object_store` is only supported for Ray Actors.
-        self.gpu_object_store: Dict[str, List["torch.Tensor"]] = {}
         # A dictionary that maps from owned object's ID to GPUObjectMeta.
+        # This dictionary is hosted on the "driver" process of the actors that
+        # store and send/receive GPU objects.
         self.managed_gpu_object_metadata: Dict[str, GPUObjectMeta] = {}
+        # Per-actor local storage for GPU objects. We create the GPU object
+        # store lazily, if a user specifies a non-default tensor_transport, to
+        # avoid circular import and because it imports third-party dependencies
+        # like PyTorch.
+        self._gpu_object_store: Optional["GPUObjectStore"] = None
 
-    def has_gpu_object(self, obj_id: str) -> bool:
-        return obj_id in self.gpu_object_store
-
-    def get_gpu_object(self, obj_id: str) -> Optional[List["torch.Tensor"]]:
-        return self.gpu_object_store[obj_id]
-
-    def add_gpu_object(self, obj_id: str, gpu_object: List["torch.Tensor"]):
-        self.gpu_object_store[obj_id] = gpu_object
-
-    def remove_gpu_object(self, obj_id: str):
-        del self.gpu_object_store[obj_id]
+    @property
+    def gpu_object_store(self) -> "ray.experimental.GPUObjectStore":
+        if self._gpu_object_store is None:
+            from ray.experimental.gpu_object_manager.gpu_object_store import GPUObjectStore
+            self._gpu_object_store = GPUObjectStore()
+        return self._gpu_object_store
 
     def _get_tensor_meta(
         self, src_actor: "ray.actor.ActorHandle", obj_id: str
@@ -148,16 +77,6 @@ class GPUObjectManager:
         # The metadata is a list of tuples, where each tuple contains the shape and dtype
         # of a tensor in the GPU object store. This function returns an ObjectRef that
         # points to the tensor metadata.
-        def __ray_get_tensor_meta__(self, obj_id: str):
-            from ray._private.worker import global_worker
-
-            gpu_object_manager = global_worker.gpu_object_manager
-            assert gpu_object_manager.has_gpu_object(
-                obj_id
-            ), f"obj_id={obj_id} not found in GPU object store"
-            tensors = gpu_object_manager.get_gpu_object(obj_id)
-            return [(t.shape, t.dtype) for t in tensors]
-
         return src_actor.__ray_call__.remote(__ray_get_tensor_meta__, obj_id)
 
     def is_managed_gpu_object(self, obj_id: str) -> bool:
@@ -188,6 +107,8 @@ class GPUObjectManager:
             src_actor: The actor that executes the task and that creates the GPU object.
             tensor_transport: The tensor transport protocol to use for the GPU object.
         """
+        from ray.experimental.gpu_object_manager.gpu_object_store import _tensor_transport_to_collective_backend
+
         tensor_transport_backend = _tensor_transport_to_collective_backend(
             tensor_transport
         )
@@ -210,6 +131,8 @@ class GPUObjectManager:
         obj_id: str,
         dst_rank: int,
     ):
+        from ray.experimental.gpu_object_manager.gpu_object_store import __ray_send__
+
         # Send tensors stored in the `src_actor`'s GPU object store to the
         # destination rank `dst_rank`.
         src_actor.__ray_call__.remote(__ray_send__, communicator_name, obj_id, dst_rank)
@@ -222,6 +145,8 @@ class GPUObjectManager:
         src_rank: int,
         tensor_meta: List[Tuple["torch.Size", "torch.dtype"]],
     ):
+        from ray.experimental.gpu_object_manager.gpu_object_store import __ray_recv__
+
         # Receive tensors from the source rank and store them in the
         # `dst_actor`'s GPU object store.
         dst_actor.__ray_call__.remote(
@@ -243,7 +168,9 @@ class GPUObjectManager:
         Returns:
             None
         """
-        if obj_id in self.gpu_object_store:
+        from ray._private.worker import global_worker
+
+        if self.gpu_object_store.has_gpu_object(obj_id):
             return
 
         gpu_object_meta = self.managed_gpu_object_metadata[obj_id]
@@ -251,7 +178,7 @@ class GPUObjectManager:
         tensors = ray.get(
             src_actor.__ray_call__.remote(__ray_fetch_gpu_object__, obj_id)
         )
-        self.gpu_object_store[obj_id] = tensors
+        self.gpu_object_store.add_gpu_object(obj_id, tensors)
 
     def trigger_out_of_band_tensor_transfer(
         self, dst_actor: "ray.actor.ActorHandle", task_args: Tuple[Any, ...]
