+++ python/ray/tests/test_gpu_objects_gloo.py
@@ -19,9 +19,9 @@ class GPUTestActor:
         return data * 2
 
     def get_gpu_object(self, obj_id: str):
-        gpu_object_manager = ray._private.worker.global_worker.gpu_object_manager
-        if gpu_object_manager.has_gpu_object(obj_id):
-            gpu_object = gpu_object_manager.get_gpu_object(obj_id)
+        gpu_object_store = ray._private.worker.global_worker.gpu_object_manager.gpu_object_store
+        if gpu_object_store.has_gpu_object(obj_id):
+            gpu_object = gpu_object_store.get_gpu_object(obj_id)
             print(f"gpu_object: {gpu_object}")
             return gpu_object
         return None
